Squashed commit of the following:
commit 02d0149f3cbb6ac96c82e61bd7100e71f91ce66d
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Wed Dec 8 22:41:07 2021 +0800

    Move all public traits used by final contract into module `interfaces`.

commit bf0a06eb26ea0257e6e54fc2511d14d7bc95bb81
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Wed Dec 8 21:23:34 2021 +0800

    Refactor implementation of validator set histories.

commit 1a37098821805273a2c77e5bb3dc296f823cf70a
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Wed Dec 8 11:45:53 2021 +0800

    Refactor implementation of staking histories.

commit 1dc4d70d8d8c4b3e8b8035a83e5320afc0b3c326
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Wed Dec 8 11:29:04 2021 +0800

    Refactor appchain notification histories and optimize testing code.

commit 425ff206d244a6d6b932c23e7d2ab5b13a4ab15d
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Wed Dec 8 09:16:32 2021 +0800

    Refactor implementation of anchor event histories and update testing code for storage migration.
