+++ appchain-anchor/src/anchor_viewer.rs
@@ -1,119 +1,4 @@
-use crate::*;
-
-pub trait AnchorViewer {
-    /// Get anchor settings detail.
-    fn get_anchor_settings(&self) -> AnchorSettings;
-    /// Get appchain settings detail.
-    fn get_appchain_settings(&self) -> AppchainSettings;
-    /// Get protocol settings detail.
-    fn get_protocol_settings(&self) -> ProtocolSettings;
-    /// Get info of OCT token.
-    fn get_oct_token(&self) -> OctToken;
-    /// Get info of wrapped appchain token.
-    fn get_wrapped_appchain_token(&self) -> WrappedAppchainToken;
-    /// Get info of near fungible tokens which has registered in this contract.
-    fn get_near_fungible_tokens(&self) -> Vec<NearFungibleToken>;
-    /// Get state of corresponding appchain.
-    fn get_appchain_state(&self) -> AppchainState;
-    /// Get current status of anchor.
-    fn get_anchor_status(&self) -> AnchorStatus;
-    /// Get validator set history info.
-    fn get_validator_set_info_of(&self, era_number: U64) -> Option<ValidatorSetInfo>;
-    /// Get processing status of validator set of era.
-    fn get_processing_status_of(&self, era_number: U64) -> Option<ValidatorSetProcessingStatus>;
-    /// Get the index range of staking histories stored in anchor.
-    fn get_index_range_of_staking_history(&self) -> IndexRange;
-    /// Get staking history by index.
-    /// If the param `index `is omitted, the latest history will be returned.
-    /// If the paran `index` is smaller than the start index, or bigger than the end index
-    /// stored in anchor, or there is no history in anchor yet, `Option::None` will be returned.
-    fn get_staking_history(&self, index: Option<U64>) -> Option<StakingHistory>;
-    /// Get the index range of anchor events stored in anchor.
-    fn get_index_range_of_anchor_event_history(&self) -> IndexRange;
-    /// Get anchor event by index.
-    /// If the param `index `is omitted, the latest event will be returned.
-    /// If the paran `index` is smaller than the start index, or bigger than the end index
-    /// stored in anchor, or there is no event in anchor yet, `Option::None` will be returned.
-    fn get_anchor_event_history(&self, index: Option<U64>) -> Option<AnchorEventHistory>;
-    /// Get anchor event by start index and quantity.
-    /// If the param `quantity` is omitted, up to 50 events will be returned.
-    fn get_anchor_event_histories(
-        &self,
-        start_index: U64,
-        quantity: Option<U64>,
-    ) -> Vec<AnchorEventHistory>;
-    /// Get the index range of appchain notification histories stored in anchor.
-    fn get_index_range_of_appchain_notification_history(&self) -> IndexRange;
-    /// Get appchain notification by index.
-    /// If the param `index `is omitted, the latest notification will be returned.
-    /// If the paran `index` is smaller than the start index, or bigger than the end index
-    /// stored in anchor, or there is no event in anchor yet, `Option::None` will be returned.
-    fn get_appchain_notification_history(
-        &self,
-        index: Option<U64>,
-    ) -> Option<AppchainNotificationHistory>;
-    /// Get appchain notification history by start index and quantity.
-    /// If the param `quantity` is omitted, up to 50 events will be returned.
-    fn get_appchain_notification_histories(
-        &self,
-        start_index: U64,
-        quantity: Option<U64>,
-    ) -> Vec<AppchainNotificationHistory>;
-    /// Get the validator list of a certain era.
-    fn get_validator_list_of(&self, era_number: Option<U64>) -> Vec<AppchainValidator>;
-    /// Get the delegators of a validator of a certain era.
-    /// If the param `era_number` is omitted, the latest validator set will be used.
-    fn get_delegators_of_validator_in_era(
-        &self,
-        era_number: Option<U64>,
-        validator_id: AccountId,
-    ) -> Vec<AppchainDelegator>;
-    /// Get unbonded stakes of an account.
-    fn get_unbonded_stakes_of(&self, account_id: AccountId) -> Vec<UnbondedStake>;
-    /// Get validator rewards of a certain era range.
-    fn get_validator_rewards_of(
-        &self,
-        start_era: U64,
-        end_era: U64,
-        validator_id: AccountId,
-    ) -> Vec<RewardHistory>;
-    /// Get validator rewards of a certain era range.
-    fn get_delegator_rewards_of(
-        &self,
-        start_era: U64,
-        end_era: U64,
-        delegator_id: AccountId,
-        validator_id: AccountId,
-    ) -> Vec<RewardHistory>;
-    /// Get current storage balance needed by this contract account.
-    fn get_storage_balance(&self) -> U128;
-    /// Get deposit of a certain validator in a certain era.
-    fn get_validator_deposit_of(&self, validator_id: AccountId, era_number: Option<U64>) -> U128;
-    /// Get deposit of a certain delegator in a certain era.
-    fn get_delegator_deposit_of(
-        &self,
-        delegator_id: AccountId,
-        validator_id: AccountId,
-        era_number: Option<U64>,
-    ) -> U128;
-    /// Get delegation list of a certain delegator in a certain era.
-    fn get_delegations_of(
-        &self,
-        delegator_id: AccountId,
-        era_number: Option<U64>,
-    ) -> Vec<AppchainDelegator>;
-    /// Get profile of a certain validator.
-    fn get_validator_profile(&self, validator_id: AccountId) -> Option<ValidatorProfile>;
-    /// Get validator profile by his/her account id in appchain.
-    fn get_validator_profile_by_id_in_appchain(
-        &self,
-        validator_id_in_appchain: String,
-    ) -> Option<ValidatorProfile>;
-    /// Get the latest commitment data of appchain state
-    fn get_latest_commitment_of_appchain(&self) -> Option<AppchainCommitment>;
-    ///
-    fn get_beefy_light_client_status(&self) -> BeefyLightClientStatus;
-}
+use crate::{interfaces::AnchorViewer, *};
 
 #[near_bindgen]
 impl AnchorViewer for AppchainAnchor {
