+++ build.sh
@@ -12,5 +12,9 @@ cp target/wasm32-unknown-unknown/release/*.wasm ./res/
 cp target/wasm32-unknown-unknown/release/appchain_anchor.wasm ./out/main.wasm
 
 if [ "$1" == "test" ]; then
-    RUST_BACKTRACE=full cargo test --test $2 -- --nocapture
+    if [ "$2" == "" ]; then
+        RUST_BACKTRACE=1 cargo test --tests -- --nocapture
+    else
+        RUST_BACKTRACE=1 cargo test --test $2 -- --nocapture
+    fi
 fi
